requirements.txt
streamlit==1.28.0
discord.py==2.3.2
pandas==2.1.2
openpyxl==3.1.2
python-dotenv==1.0.0
asyncio 